+++ CHANGELOG.rst
@@ -1,3 +1,8 @@
+100.2.3 (2026-04-24)
+--------------------
+* Removed RestrictionCapableEval in order to prep for Zope 6
+* Removed startZServer from tests/utils
+
 100.2.2 (2026-04-22)
 --------------------
 * Fixed bdr listing for envelopes without title in engineMacros